Completed rest api for routes and added user auth

+++ backend/controllers/workoutController.js
@@ -1,11 +1,14 @@
 const mongoose = require("mongoose");
 const Workout = require("../models/workoutModel");
+const User = require("../models/userModel");
 
 // @desc    Get all workouts
 // @route   GET /api/workouts
 // @access  Private
 const getAllWorkouts = async (req, res) => {
-  const retrievedWorkouts = await Workout.find({}).sort({ createdAt: -1 });
+  const retrievedWorkouts = await Workout.find({ user: req.user.id }).sort({
+    createdAt: -1,
+  });
   res.status(200).json(retrievedWorkouts);
 };
 
@@ -16,7 +19,12 @@ const getWorkout = async (req, res) => {
   const { id } = req.params;
 
   validIdCheck(id, res);
-  const retrievedWorkout = await Workout.findById(id);
+
+  const retrievedWorkout = await Workout.findOne({
+    _id: id,
+    user: req.user.id,
+  });
+
   returnStatus(retrievedWorkout, res);
 };
 
@@ -25,9 +33,8 @@ const getWorkout = async (req, res) => {
 // @access  Private
 const postWorkout = async (req, res) => {
   try {
-    const body = req.body;
-    const obj = new Workout(body);
-    const postedWorkout = await Workout.create(obj);
+    const body = { user: req.user.id, ...req.body };
+    const postedWorkout = await Workout.create(new Workout(body));
     res.status(200).json(postedWorkout);
   } catch (error) {
     res.status(400).json({ error: error.message });
@@ -42,8 +49,28 @@ const updateWorkout = async (req, res) => {
   const body = req.body;
 
   validIdCheck(id, res);
+
+  //find and verify workout
+  const workout = await Workout.findById(id);
+
+  if (!workout) {
+    return res.status(400).json({ error: "Workout not found." });
+  }
+
+  //find and verify user
+  const user = await User.findById(req.user.id);
+
+  if (!user) {
+    return res.status(400).json({ error: "User not found." });
+  }
+
+  //check if the logged in user matches the user attached to the specific workout
+  if (user.id !== workout.user.toString()) {
+    return res.status(401).json({ error: "User not authorized." });
+  }
+
   const updatedWorkout = await Workout.findByIdAndUpdate(id, body);
-  returnStatus(updatedWorkout, res);
+  res.status(200).json(updatedWorkout);
 };
 
 // @desc    Delete a workout
@@ -53,8 +80,26 @@ const deleteWorkout = async (req, res) => {
   const { id } = req.params;
 
   validIdCheck(id, res);
+
+  //find and verify workout
+  const workout = await Workout.findById(id);
+  if (!workout) {
+    return res.status(400).json({ error: "Workout not found" });
+  }
+
+  // find and verify user
+  const user = await User.findById(req.user.id);
+  if (!user) {
+    return res.status(400).json({ error: "User not found." });
+  }
+
+  //check if the logged in user matches the user attached to the specific workout
+  if (user.id !== workout.user.toString()) {
+    return res.status(401).json({ error: "User not authorized." });
+  }
+
   const deletedWorkout = await Workout.findByIdAndDelete(id);
-  returnStatus(deletedWorkout, res);
+  res.status(200).json(deletedWorkout);
 };
 
 //helper functions
